+++ Source/SIMPLib/Geometry/ImageGeom.cpp
@@ -92,7 +92,7 @@ class FindImageDerivativesImpl
       size_t dims[3] = { 0, 0, 0 };
       m_Image->getDimensions(dims);
 
-      size_t counter = 0;
+      int64_t counter = 0;
       size_t totalElements = m_Image->getNumberOfElements();
       int64_t progIncrement = static_cast<int64_t>(totalElements / 100);
